# desafioFinalGo/docs/swagger.yaml
definitions:
...
  domain.Paciente:
    properties:
      data_cadastro:
        type: string
      id:
        type: integer
      nome:
        type: string
      rg:
        type: integer
      sorenome:
        type: string
    type: object
  web.errorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact:
    name: Public APIs
    url: https://dev.publicapis.ml
  description: This API handle products from our store
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://publicapis.ml/terms
  title: Store API
  version: "1.0"
paths:
...
  /pacientes:
    get:
      consumes:
      - application/json
      description: get pacientes
      parameters:
      - description: SECRET_TOKEN
        in: header
        name: SECRET_TOKEN
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Paciente'
      summary: List pacientes
      tags:
      - CRUD PACIENTES
    post:
      consumes:
      - application/json
      description: Post a paciente
      parameters:
      - description: SECRET_TOKEN
        in: header
        name: SECRET_TOKEN
        required: true
        type: string
      - description: Paciente
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/domain.Paciente'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Paciente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Post a paciente
      tags:
      - CRUD PACIENTES
  /pacientes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a paciente by ID
      parameters:
      - description: SECRET_TOKEN
        in: header
        name: SECRET_TOKEN
        required: true
        type: string
      - description: Paciente ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Paciente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Delete a paciente by ID
      tags:
      - CRUD PACIENTES
    get:
      consumes:
      - application/json
      description: Get a paciente by ID
      parameters:
      - description: SECRET_TOKEN
        in: header
        name: SECRET_TOKEN
        required: true
        type: string
      - description: Paciente ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Paciente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Get a paciente by ID
      tags:
      - CRUD PACIENTES
    patch:
      consumes:
      - application/json
      description: Patch a paciente
      parameters:
      - description: SECRET_TOKEN
        in: header
        name: SECRET_TOKEN
        required: true
        type: string
      - description: Paciente ID
        in: path
        name: id
        required: true
        type: integer
      - description: Paciente
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/domain.Paciente'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Paciente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Patch a paciente
      tags:
      - CRUD PACIENTES
    put:
      consumes:
      - application/json
      description: Put a paciente
      parameters:
      - description: SECRET_TOKEN
        in: header
        name: SECRET_TOKEN
        required: true
        type: string
      - description: Paciente ID
        in: path
        name: id
        required: true
        type: integer
      - description: Paciente
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/domain.Paciente'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Paciente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Put a paciente
      tags:
      - CRUD PACIENTES
swagger: "2.0"